Use Long instead of int for IDs. Sql Connection encoding UTF8

+++ main/java/slugchat/mobile/service/dao/ClientDao.java
@@ -4,7 +4,7 @@ import main.java.slugchat.mobile.service.domain.Client;
 
 public interface ClientDao {
 
-    Client getClient(int clientId);
+    Client getClient(Long clientId);
     Client createClient(Client client);
 
 }
